==> docs/notebooks/02_api_vs_cli.ipynb <==
# Command Line Interface (CLI)
IsoTools is a python library for the analysis of long read data. 
It starts from aligned reads (e.g. bamfiles) and features transcriptome reconstruction, quantification, visualization, explorative analysis and statistical tests.
IsoTools is designed to be run interactively (e.g. in a python notebook, or in the repl), as it features rich functionality to explore the data. 
However, it can also be run as a command line tool, with predefined parameters. This is convenient to process different experiments in a standardized pipeline.
All command line parameters are described in the [CLI section](../isotoolsCLI.html):


To run the commands, need the following **input files** from the [prepared demonstration dataset](https://nc.molgen.mpg.de/cloud/index.php/s/zYe7g6qnyxGDxRd):

* sample description file 'encode_samples.tsv'
* six .bam alignment files for the six samples
* six corresponding .bam.bai indices
* reference annotation file gencode.v42.chr_patch_hapl_scaff.annotation_sorted_chr8.gff3.gz
* corresponding .gff3.gz.tbi index file
* genomic reference file GRCh38.p13.genome_chr8.fa
* corresponding .fai index file

The sample description file is a tab-separated text file with at least the following columns:

* **sample_name**: A unique sample label
* **file_name**: The path to the alignment .bam file
* **group**: a group assignment of the sample


The following command runs IsoTools and performs the following analysis steps:

* transcriptome reconstruction from the alignments
* filtering transcripts according to filter query (--filter_query)
    * see [filter tutorial](06_filtering.html#Filtering-tags-and-queries) for a description of tags and queries.
* exporting transcripts as gtf (--gtf_out)
* exporting a transcript table, which includes the number of reads per transcripts (--transcript_table)
* finally, by default, a pkl file is produced, which contains the transcriptome object, reducing runtime of the next call of the CLI.
    * The resulting transcriptome can also be explored using the API.
    * creation of the pkl file can be skipped with --no-pkl
    * import of the previous pkl file can be bypassed with the --force-recreate command
    * if a pkl file is present, and a sample table containing novel samples is provided, only the novel samples are added
* the two parameters "--log" INFO and "--progress_bar" enable readout of the progress. 

%%bash
mkdir -p demonstration_dataset
cd demonstration_dataset
samples='encode_samples.tsv'
anno='gencode.v42.chr_patch_hapl_scaff.annotation_sorted_chr8.gff3.gz'
genome='GRCh38.p13.genome_chr8.fa'

run_isotools \
    --anno $anno \
    --log INFO \
    --progress_bar \
    --genome $genome \
    --samples $samples \
    --file_prefix ./PacBio_isotools_substantial \
    --custom_filter_tag "COVERED=any(gene.coverage[:,transcript_id] > 2)"  "HIGH_COVER=gene.coverage.sum(0)[transcript_id] >= 7" \
    --filter_query "(COVERED and FSM) or (HIGH_COVER and SUBSTANTIAL and not INTERNAL_PRIMING)" \
    --gtf_out --transcript_table

Running the command for the next time will automatically load the stored '.pkl' file. 
Here, we use the command line interface to perform differential splicing analysis.
The resulting table with test statistics for all alternative splicing events is written to the file 
PacBio_isotools_substantial_diff_K562_GM12878.csv.
In addidtion, the command creates sashimi coverage plots for the top 5 differentially spliced genes

%%bash
cd demonstration_dataset
run_isotools \
    --log INFO \
    --progress_bar \
    --file_prefix ./PacBio_isotools_substantial \
    --group_by group \
    --diff_plots 5 \
    --diff K562/GM12878
